python_service/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-aiplatform
ffmpeg-python
requests
python-multipart
python-dotenv
motor
google-auth
google-auth-httplib2
stripe
stripe
pydantic
cryptography
